+++ core/src/main/java/org/eqasim/core/analysis/DistanceUnit.java
@@ -1,0 +1,5 @@
+package org.eqasim.core.analysis;
+
+public enum DistanceUnit {
+	kilometer, meter, mile, foot
+}