+++ core/src/main/java/org/eqasim/core/simulation/mode_choice/AbstractEqasimExtension.java
@@ -1,8 +1,12 @@
 package org.eqasim.core.simulation.mode_choice;
 
+import java.util.Map;
+
+import org.eqasim.core.components.config.EqasimConfigGroup;
 import org.eqasim.core.simulation.mode_choice.cost.CostModel;
 import org.eqasim.core.simulation.mode_choice.utilities.UtilityEstimator;
 
+import com.google.inject.Provider;
 import com.google.inject.binder.LinkedBindingBuilder;
 import com.google.inject.multibindings.MapBinder;
 
@@ -29,4 +33,21 @@ public abstract class AbstractEqasimExtension extends AbstractDiscreteModeChoice
 	}
 
 	abstract protected void installEqasimExtension();
+
+	protected CostModel getCostModel(Map<String, Provider<CostModel>> factory, EqasimConfigGroup config, String mode) {
+		String model = config.getCostModels().get(mode);
+
+		if (model == null) {
+			throw new IllegalStateException(String.format("No cost model defined for mode '%s'", mode));
+		} else {
+			Provider<CostModel> modelFactory = factory.get(model);
+
+			if (modelFactory == null) {
+				throw new IllegalStateException(
+						String.format("Cost model '%s' for mode '%s' is not known", model, mode));
+			} else {
+				return modelFactory.get();
+			}
+		}
+	}
 }